Allow properties when creating a basejail
The iocage command line tool allows us to set properties when creating a basejail, and they work. In fact, in order to 1) create a basejail and 2) install packages, you must set a network property (ip4_addr or ip6_addr) in order for pkg to be able to retrieve said packages. But you cannot currently do this in the iocage Ansible module, since even if you set properties, setting the module's state to "basejail" causes it to overwrite any set properties. This change allows through the properties set by the user.